Move logic of dispatch into performOperation function
